+++ addons/website_crm/__manifest__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 {
     'name': 'Contact Form',
     'category': 'Website/Website',
@@ -14,11 +15,12 @@ This module includes contact phone and mobile numbers validation.""",
     'depends': ['website', 'crm'],
     'data': [
         'security/ir.model.access.csv',
-        'data/website_crm_data.xml',
-        'views/website_crm_lead_views.xml',
-        'views/website_crm_templates.xml',
+        'data/ir_actions_data.xml',
+        'data/ir_model_data.xml',
+        'views/crm_lead_views.xml',
         'views/res_config_settings_views.xml',
         'views/website_visitor_views.xml',
+        'views/website_templates_contactus.xml',
     ],
     'installable': True,
     'auto_install': True,
