+++ addons/website_crm/data/ir_model_data.xml
@@ -7,6 +7,7 @@
             <field name="website_form_access">True</field>
             <field name="website_form_label">Create an Opportunity</field>
         </record>
+
         <function model="ir.model.fields" name="formbuilder_whitelist">
             <value>crm.lead</value>
             <value eval="[
@@ -20,17 +21,5 @@
                 'user_id',
             ]"/>
         </function>
-    </data>
-    <data noupdate="1">
-        <record id="action_open_website" model="ir.actions.act_url">
-            <field name="name">Website Contact Form</field>
-            <field name="target">self</field>
-            <field name="url">/contactus</field>
-        </record>
-
-        <record id="base.open_menu" model="ir.actions.todo">
-            <field name="action_id" ref="action_open_website"/>
-            <field name="state">open</field>
-        </record>
     </data>
 </odoo>
