+++ addons/website_crm/models/__init__.py
@@ -3,4 +3,5 @@
 
 from . import crm_lead
 from . import res_config_settings
+from . import website
 from . import website_visitor
