+++ addons/website_crm/models/crm_lead.py
@@ -53,21 +53,3 @@ class Lead(models.Model):
                             request.website.crm_default_user_id.id
         values['type'] = 'lead' if self.with_user(SUPERUSER_ID).env['res.users'].has_group('crm.group_use_lead') else 'opportunity'
         return values
-
-
-class Website(models.Model):
-    _inherit = 'website'
-
-    def _get_crm_default_team_domain(self):
-        if not self.env.user.has_group('crm.group_use_lead'):
-            return [('use_opportunities', '=', True)]
-        return [('use_leads', '=', True)]
-
-    crm_default_team_id = fields.Many2one(
-        'crm.team', string='Default Sales Teams',
-        default=lambda self: self.env['crm.team'].search([], limit=1),
-        domain=lambda self: self._get_crm_default_team_domain(),
-        help='Default Sales Team for new leads created through the Contact Us form.')
-    crm_default_user_id = fields.Many2one(
-        'res.users', string='Default Salesperson', domain=[('share', '=', False)],
-        help='Default salesperson for new leads created through the Contact Us form.')
