+++ addons/website_crm/views/website_templates_contactus.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
+    <data>
         <template id="contactus_form" name="Contact Form (Opportunity)" inherit_id="website.contactus" customize_show="True">
             <xpath expr="//span[@data-for='contactus_form']" position="before">
                 <t t-set="contactus_form_values" t-value="dict(contactus_form_values, **{
@@ -12,5 +13,5 @@
                 })"/>
             </xpath>
 		</template>
-
-        </odoo>
+    </data>
+</odoo>
