+++ addons/website_crm/views/website_visitor_views.xml
@@ -1,26 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo><data>
-    <!--Website visitor actions-->
-    <record id="website_crm.website_visitor_crm_lead_action" model="ir.actions.act_window">
-        <field name="name">Leads</field>
-        <field name="res_model">crm.lead</field>
-        <field name="view_mode">tree,form</field>
-        <field name="domain">[('visitor_ids', 'in', [active_id])]</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_empty_folder">
-              No lead linked for this visitor
-            </p>
-        </field>
-    </record>
-
-    <!-- website visitor views -->
     <record id="website_visitor_view_form" model="ir.ui.view">
         <field name="name">website.visitor.view.form.inherit.website.crm</field>
         <field name="model">website.visitor</field>
         <field name="inherit_id" ref="website.website_visitor_view_form"/>
         <field name="arch" type="xml">
             <xpath expr="//button[@id='w_visitor_visit_counter']" position="before">
-                <button name="%(website_crm.website_visitor_crm_lead_action)d" type="action" class="oe_stat_button" icon="fa-star"
+                <button name="%(website_crm.crm_lead_action_from_visitor)d" type="action" class="oe_stat_button" icon="fa-star"
                         attrs="{'invisible': [('lead_count', '=', 0)]}">
                     <field name="lead_count" widget="statinfo" string="Leads"/>
                 </button>
